# web/frontend-v2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Z-Image Generator",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/models/{pipeline_id}/load": {
      "post": {
        "summary": "Load Model By Id",
        "description": "Load a model by pipeline ID.\n\nThis is the unified API for the React frontend. Maps pipeline IDs to\nthe specific load functions, passing through DI dependencies.",
        "operationId": "load_model_by_id_api_models__pipeline_id__load_post",
        "parameters": [
          {
            "name": "pipeline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pipeline Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessVramResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{pipeline_id}/unload": {
      "post": {
        "summary": "Unload Model By Id",
        "description": "Unload a model by pipeline ID.",
        "operationId": "unload_model_by_id_api_models__pipeline_id__unload_post",
        "parameters": [
          {
            "name": "pipeline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pipeline Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessVramResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/models/{pipeline_id}/status": {
      "get": {
        "summary": "Get Model Status",
        "description": "Get the status of a specific pipeline model.",
        "operationId": "get_model_status_api_models__pipeline_id__status_get",
        "parameters": [
          {
            "name": "pipeline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pipeline Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/loras/{pipeline_id}": {
      "get": {
        "summary": "List Loras For Pipeline",
        "description": "List LoRA files available for a specific pipeline.",
        "operationId": "list_loras_for_pipeline_api_loras__pipeline_id__get",
        "parameters": [
          {
            "name": "pipeline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pipeline Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoRAListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigTag": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "color": {
            "type": "string",
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "color"
        ],
        "title": "ConfigTag"
      },
      "ConfigWarning": {
        "properties": {
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "severity",
          "message"
        ],
        "title": "ConfigWarning"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoRAFileInfo": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "directory": {
            "type": "string",
            "title": "Directory"
          },
          "sizeMb": {
            "type": "number",
            "title": "Sizemb"
          }
        },
        "type": "object",
        "required": [
          "path",
          "name",
          "directory",
          "sizeMb"
        ],
        "title": "LoRAFileInfo",
        "description": "A single LoRA file on disk."
      },
      "LoRAInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "scale": {
            "type": "number",
            "title": "Scale"
          },
          "layersUpdated": {
            "type": "integer",
            "title": "Layersupdated"
          }
        },
        "type": "object",
        "required": [
          "name",
          "path",
          "scale",
          "layersUpdated"
        ],
        "title": "LoRAInfo",
        "description": "LoRA fusion state for a single LoRA applied to the active model."
      },
      "LoRAListResponse": {
        "properties": {
          "loras": {
            "items": {
              "$ref": "#/components/schemas/LoRAFileInfo"
            },
            "type": "array",
            "title": "Loras"
          },
          "directories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Directories"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "default": 0
          },
          "pipelineId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipelineid"
          }
        },
        "type": "object",
        "title": "LoRAListResponse",
        "description": "Available LoRA files from /api/loras."
      },
      "ModelStatusResponse": {
        "properties": {
          "pipelineId": {
            "type": "string",
            "title": "Pipelineid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "components": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Components"
          },
          "totalVramMb": {
            "type": "integer",
            "title": "Totalvrammb",
            "default": 0
          },
          "vramMb": {
            "type": "integer",
            "title": "Vrammb",
            "default": 0
          },
          "modelVariant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelvariant"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "loras": {
            "items": {
              "$ref": "#/components/schemas/LoRAInfo"
            },
            "type": "array",
            "title": "Loras"
          },
          "loraSummary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lorasummary"
          },
          "configTags": {
            "items": {
              "$ref": "#/components/schemas/ConfigTag"
            },
            "type": "array",
            "title": "Configtags"
          },
          "configWarnings": {
            "items": {
              "$ref": "#/components/schemas/ConfigWarning"
            },
            "type": "array",
            "title": "Configwarnings"
          }
        },
        "type": "object",
        "required": [
          "pipelineId",
          "status"
        ],
        "title": "ModelStatusResponse",
        "description": "Status for a single pipeline model from /api/models/{id}/status."
      },
...
      "SuccessVramResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "vram": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vram"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "SuccessVramResponse",
        "description": "Success response with optional VRAM snapshot."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}